Fix uint issues

+++ src/model/food_bag.rs
@@ -2,12 +2,12 @@ use crate::model::food::Food;
 
 pub struct FoodBag {
     items: Vec<Food>,
-    total_calories: i32,
+    total_calories: u16,
 }
 
 impl FoodBag {
     pub fn new(items: Vec<Food>) -> Self {
-        let mut total_calories: u32 = 0;
+        let mut total_calories: u16 = 0;
         for food in items.iter() {
             total_calories = total_calories + food.calories;
         }
